refactor(ios): split native auth exchange out of APIClient (#14267)
Behavior-preserving move (APIClient.swift 463 -> 359 lines): the
native auth exchange unit (NativeAuthExchangeError,
NativeAuthExchangeResponse, NativeAuthExchangeClient, and the private
request payload) moves verbatim to Core/AuthTokenExchange.swift.
APIClientError stays with APIClient. pbxproj wired with explicit
entries.
